Slight reorganization.
