+++ crates/dprint/src/environment/real_environment.rs
@@ -234,7 +234,6 @@ impl Environment for RealEnvironment {
     }
 }
 
-
 struct CustomDialoguerTheme {
 }
