+++ crates/dprint/src/plugins/cache_manifest.rs
