+++ crates/dprint/src/plugins/mod.rs
@@ -1,21 +1,19 @@
 pub mod wasm;
-pub mod process;
+mod process;
 mod plugin;
-mod plugin_cache;
-mod plugin_cache_manifest;
-mod plugin_source_reference;
-mod plugin_resolver;
+mod cache;
+mod cache_manifest;
+mod common;
+mod resolver;
 mod pool;
-mod setup_plugin;
 mod repo;
 mod types;
 
 pub use plugin::*;
-pub use plugin_cache::*;
-use plugin_cache_manifest::*;
-pub use setup_plugin::*;
-pub use plugin_source_reference::*;
-pub use plugin_resolver::*;
+pub use cache::*;
+use common::*;
+use cache_manifest::*;
+pub use resolver::*;
 pub use pool::*;
 pub use repo::*;
 pub use types::*;
